Organize fragments and page query
