+++ apps/server/src/presentation/hello/hello.graphql
