+++ apps/server/src/presentation/items/items.graphql
