+++ apps/web/README.md
@@ -17,3 +17,11 @@ npm ci
 
 npm run dev
 ```
+
+### Hello Page
+
+The `Hello` component fetches its message from the GraphQL server using Apollo
+Client. The fragment is defined alongside the component in
+`components/Hello/fragment.ts` and a page-level query assembling those fragments
+lives in `templates/hello.ts`. The container executes that query with
+`useQuery`.
