+++ apps/web/components/Hello/fragment.ts
@@ -1,0 +1,7 @@
+import { gql } from '@apollo/client';
+
+export const HELLO_MESSAGE = gql`
+  fragment HelloMessage on Query {
+    hello
+  }
+`;
