+++ apps/web/components/Hello/index.tsx
@@ -1,6 +1,18 @@
+import { useQuery } from '@apollo/client';
 import { HelloPresentation } from './components';
+import { HELLO_QUERY } from '../../templates/hello';
 
 export default function HelloContainer() {
-  const message = 'Hello from Next.js 14 Page Router!';
+  const { data, loading, error } = useQuery(HELLO_QUERY);
+
+  let message = 'Loading...';
+  if (loading) {
+    message = 'Loading...';
+  } else if (error) {
+    message = 'Error fetching data';
+  } else {
+    message = data?.hello ?? 'No data';
+  }
+
   return <HelloPresentation message={message} />;
 }
