+++ apps/web/package.json
@@ -9,6 +9,7 @@
     "codegen": "graphql-codegen"
   },
   "dependencies": {
+    "@apollo/client": "^3.8.0",
     "next": "14.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
