+++ apps/web/utils/apolloClient.ts
@@ -1,0 +1,6 @@
+import { ApolloClient, InMemoryCache, HttpLink } from '@apollo/client';
+
+export const client = new ApolloClient({
+  link: new HttpLink({ uri: 'http://localhost:3001/graphql' }),
+  cache: new InMemoryCache(),
+});
